refactor(bench): separate wall-time runner support
